Advance prologue gameplay and native rendering

+++ game/include/main/dll/expgfx_interface.h
@@ -6,10 +6,10 @@
 typedef void (*ExpgfxOnMapSetupFn)(void);
 typedef int (*ExpgfxSpawnEffectFn)(void* config, int preferredPoolIndex, int sourceId,
                                    int flags);
-typedef void (*ExpgfxUpdateFrameStateFn)(int sourceMode, int sourceId, int unused0,
+typedef void (*ExpgfxUpdateFrameStateFn)(int sourceMode, uintptr_t sourceId, int unused0,
                                          int unused1);
 typedef void (*ExpgfxResetAllPoolsFn)(void);
-typedef void (*ExpgfxFreeSourceFn)(u32 sourceId);
+typedef void (*ExpgfxFreeSourceFn)(uintptr_t sourceId);
 typedef int (*ExpgfxFunc09Fn)(void);
 typedef void (*ExpgfxNopFn)(void);
 typedef void (*ExpgfxUpdateSourceFrameFlagsFn)(void* sourceObject);
